seedbs

+++ seedbs.R
@@ -129,18 +129,23 @@ template_index <- as.integer(args[1])  # Start from 1
 prompt_index <- as.integer(args[2])  # Start from 0
 seeded_interval_index <- as.integer(args[3])  # Start from 1
 
-pvalue_vector <- rep(
-  NA,
-  seeded_intervals[seeded_interval_index, 2] -
-    seeded_intervals[seeded_interval_index, 1] + 1
-)
-for (i in seq_len(length(pvalue_vector))) {
-  filename <- sub("XXX", prompt_index, pvalue_files_templates[template_index])
-  filename <-
-    sub("YYY", seeded_intervals[seeded_interval_index, 1] + i - 1 - 1, filename)
-  pvalue_vector[i] <- unlist(read.csv(filename, header = FALSE))
-}
-index_p_tilde <- segment_significance(pvalue_vector)
 filename <- sub("XXX", prompt_index, pvalue_files_templates[template_index])
 filename <- sub("YYY", paste0("SeedBS-", seeded_interval_index), filename)
-write.csv(index_p_tilde, filename, row.names = FALSE)
+if (!file.exists(filename)) {
+  pvalue_vector <- rep(
+    NA,
+    seeded_intervals[seeded_interval_index, 2] -
+      seeded_intervals[seeded_interval_index, 1] + 1
+  )
+  for (i in seq_len(length(pvalue_vector))) {
+    filename <- sub("XXX", prompt_index, pvalue_files_templates[template_index])
+    filename <- sub(
+      "YYY", seeded_intervals[seeded_interval_index, 1] + i - 1 - 1, filename
+    )
+    pvalue_vector[i] <- unlist(read.csv(filename, header = FALSE))
+  }
+  index_p_tilde <- segment_significance(pvalue_vector)
+  filename <- sub("XXX", prompt_index, pvalue_files_templates[template_index])
+  filename <- sub("YYY", paste0("SeedBS-", seeded_interval_index), filename)
+  write.csv(index_p_tilde, filename, row.names = FALSE)
+}
